Make effectiveDevicePixelRatio() return a qreal.
Fractional scale factors are not as broken as previously
believed, especially for Qt Quick. Keep the door open
for finding a way to support it at some point in the
future.

+++ src/quick/items/qquickwindow.cpp
@@ -3945,7 +3945,7 @@ void QQuickWindow::runJobsAfterSwap()
  *
  * \sa QWindow::devicePixelRatio()
  */
-int QQuickWindow::effectiveDevicePixelRatio() const
+qreal QQuickWindow::effectiveDevicePixelRatio() const
 {
     QWindow *w = QQuickRenderControl::renderWindowFor(const_cast<QQuickWindow *>(this));
     return w ? w->devicePixelRatio() : devicePixelRatio();
